requirements.txt
torch
torchvision
Pillow
tqdm
numpy
